requirements.txt
requests==2.20.0
baidu_aip==2.2.17.0
pytest==4.3.0
selenium==2.53.1
uiautomator2==0.3.3